requirements.txt
astroid==1.2.1
Babel==1.3
coverage==3.7.1
django-extensions==1.4.6
Django==1.7.8
docutils==0.12
flake8-todo==0.3
flake8==2.2.5
ipython==2.3.1
logilab-common==0.63.0
MarkupSafe==0.23
mccabe==0.2.1
pip-tools==0.3.5
pycups==1.9.68
pyflakes==0.8.1
Pygments==2.0.1
pygraphviz==1.3rc2
pylint==1.3.1
python-memcached==1.53
python-mpd2==0.5.3
pytz==2014.9
PyYAML==3.11
simplejson==3.6.5
six==1.8.0
snowballstemmer==1.2.0
sphinx-bootstrap-theme==0.4.3
Sphinx==1.3b1
sqlparse==0.1.13
Werkzeug==0.9.6
